+++ CHANGELOG.md
@@ -3,4 +3,17 @@
 # AutomaticMapperGenerator Changelog
 
 ## [Unreleased]
+### Added
+- User notification when plugin finished
+- Upgraded to support only Intellij version 2021.* to avoid memory leak issue, see `readMe.md` for more info
+
+## [0.2.0] - 2021-07-01
+### Added
+- Generation of Mappers
+- Better handling of user settings
+
+## [0.1.0] - 2021-06-29
+### Added
+- Generation of Incoming Message and Big Query Schema
+- Initial UI
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)